requirements.txt
astroid==2.0.4
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
decorator==4.3.0
future==0.17.0
idna==2.7
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.15.3
oauthlib==2.1.0
pandas==0.23.4
parso==0.3.1
pickleshare==0.7.5
prompt-toolkit==2.0.7
Pygments==2.2.0
pylint==2.1.1
python-dateutil==2.7.5
python-twitter==3.4.2
pytz==2018.7
requests==2.20.0
requests-oauthlib==1.0.0
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
stop-words==2018.7.23
traitlets==4.3.2
urllib3==1.24
wcwidth==0.1.7
wrapt==1.10.11
